Make recount resources public and set content type to json

+++ buildSrc/src/main/groovy/au/id/tmm/ausvotes/buildsrc/DeleteS3Task.groovy
@@ -1,10 +1,10 @@
 package au.id.tmm.ausvotes.buildsrc
 
-import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.AmazonS3ClientBuilder
 import com.amazonaws.services.s3.model.DeleteObjectsRequest
 import com.amazonaws.services.s3.model.ObjectListing
 import org.gradle.api.DefaultTask
+import org.gradle.api.logging.LogLevel
 import org.gradle.api.tasks.Input
 import org.gradle.api.tasks.TaskAction
 
@@ -24,9 +24,15 @@ class DeleteS3Task extends DefaultTask {
 
         List<String> keys = objects.objectSummaries.collect { it.key }
 
-        def deleteObjectsRequest = new DeleteObjectsRequest(bucket)
+        if (!keys.empty) {
+            logger.log(LogLevel.INFO, "Attempting to delete keys from s3: ${keys.toListString(100)}")
+
+            def deleteObjectsRequest = new DeleteObjectsRequest(bucket)
                 .withKeys(*keys)
 
-        client.deleteObjects(deleteObjectsRequest)
+            client.deleteObjects(deleteObjectsRequest)
+        } else {
+            logger.log(LogLevel.INFO, "No action as prefix $prefix does not match any objects")
+        }
     }
 }
